Split out login and environment configuration
