+++ cli/src/commands/cloud/mod.rs
@@ -8,6 +8,7 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0.
 
+mod environments;
 mod login;
 
 use cling::prelude::*;
@@ -56,4 +57,7 @@ impl Default for CloudConfig {
 pub enum Cloud {
     /// Authenticate to Restate Cloud
     Login(login::Login),
+    /// Manage Restate Cloud Environments
+    #[clap(subcommand)]
+    Environments(environments::Environments),
 }
